Moved static files and switched save game button to patch
