+++ src/index.js
@@ -14,8 +14,8 @@ import Signup from './pages/Signup';
 import Login from './pages/Login';
 import LoginHome from './pages/LoginHome';
 import CreateNewGame from './pages/CreateNewGame';
-import GameEdit682 from './pages/GameEdit682';
-import GameEditNew from './pages/GameEditNew';
+import GameEdit682 from './pages/static/GameEdit682';
+import GameEditNew from './pages/static/GameEditNew';
 import PlayDetails from './pages/PlayDetails';
 import PlayGame from './pages/static/PlayGame';
 import MakeGames from './pages/MakeGames';
