+++ src/pages/static/GameEdit682.jsx
