+++ src/pages/static/GameEditNew.jsx
